LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entries-after-blank-navigation-from-cross-origin-expected.txt



FAIL entries() should not be leaked from cross-origin when navigating to about:blank assert_equals: expected 1 but got 0
